+++ inst/doc/SCAN_25_SCAN_100_SCAN_OACI.R
@@ -6,6 +6,7 @@ knitr::opts_chunk$set(
 
 ## ----setup--------------------------------------------------------------------
 library(happign)
+library(sf)
 
 ## ---- eval = FALSE, echo = TRUE-----------------------------------------------
 #  # Create or import a shape
@@ -22,7 +23,7 @@ library(happign)
 #  scan25 <- get_wms_raster(shape,
 #                           apikey = my_key,
 #                           layer_name = scan25_name,
-#                           resolution = NULL,   # To have full resolution
+#                           resolution = 1,   # To have full resolution
 #                           filename = "SCAN25") # This resource need to be downloaded on disk
 #  
 #