+++ tests/testthat/setup.R
@@ -1,0 +1,3 @@
+library(httptest2)
+library(stars)
+library(sf)